Reorganize packages that share result+error types
